refactor(year-filter): Moved the dashboard specific constants to separate file
