+++ apps/website/src/constants/dashboard.ts
@@ -1,0 +1,9 @@
+import { ChartConfig } from "@/components/ui/chart";
+
+export const chartConfig = {
+  common: { label: "Common", color: "#a3a3a3" },
+  rare: { label: "Rare", color: "#93c5fd" },
+  epic: { label: "Epic", color: "#a855f7" },
+  legendary: { label: "Legendary", color: "#facc15" },
+  count: { label: "Opened", color: "var(--primary)" },
+} satisfies ChartConfig;
